refactor: move area secondary/badge logic to AreaCard
